Only validate constructor javadoc when public or protected.
Tricky point! Enums should have constructors (where necessary) that
appear to be package visibility, yet they are actually private. They are
a special case in the Java Language Specification. That means that
checkstyle complains if you put an explicit “private” in, but it also
complains about missing javadoc (which is unnecessary) if you leave
“private” out. This is really a bug in checkstyle, but we need to work
around it anyway…