Fixes fast data in.
This stops the code that sets the tag for fast data in communications
from hammering SCAMP (which was happening because I'd mis-documented the
SCPReceiver implementation, which was a copy-and-paste-from-Python
error). That hammering was pushing SCAMP into a strange state, and was
triggering the later failures.

Also copy some more constant naming from Python and how to reset boards'
timeout schemes.