Get rid of the weird socket close/open shuffle.
UDP sockets simply don't have the sort of state to justify this sort of
thing! In-flight packets will simply get delivered to the new socket
instead of the old one. So why close and reopen? It's purely a piece of
prophylactic code that actually merely introduces new failure modes
without helping.