Make IPTag useSender actually work right.
When using the useSender flag of the IPTagSet operation, you
categorically need to send the message on the connection that you wish
to receive messages on because SCAMP uses both the IP address and the
port of the sending connection. Because of that, there is no point in
setting the flag except when you know what connection will be receiving,
and the API should therefore take the connection that will be doing the
receiving at that point (which needs to be an SDPConnection, but no
more).

fixes #246