Change way that requests are sent
Removes bad old decision to use number of arguments to describe what
sort of job to create. It's much better to be more explicit; this allows
us to allocate a specific board by logical, physical or network
coordinates. It might eventually allow us to specify other things… once
we think of them.