Transactions need to be mode DEFERRED, not mode IMMEDIATE
In theory this is Bad, as it could mean that transactions fail after
being entered. In practice this is Good, because it means that the
complex startup sequence of the application doesn't lock the database
file in a weird way, causing exceptions all over that I've no idea how
to fix.

It helps that we're very careful now to keep transactions short and
non-blocking.