Fix the weird transaction errors
Having read about how Connection.setAutoCommit() works, it seems that
the easiest approach is to just turn it off permanently on every
connection and do the transaction management myself. I was already
basically doing that (except in tests) so making this a permanent
condition of the connections immensely simplifies things.

Also added warnings so that running SQL outside a transaction will get
noticed.