Read locks are fine. Upgrading locks are what cause the problem.
The issue is that if you have several read-locked transactions that all
try to upgrade themselves to write-locked at once, you get a deadlock
(and eventually a timeout). This change forces transactions to say what
sort of lock they require, defaulting to (exclusive) write. There's
quite a few places where we only need read locks, which are much faster.