Get the OpenID auth realm right
Of course, this immediately failed because Java doesn't trust modern
certificates by default. This should be easy to fix, yes? No!
https://github.com/spring-projects/spring-security/issues/8882
Instead we have to fabricobble something together from
https://stackoverflow.com/a/24561444/301832
and my knowledge of how Spring works. It's at least as gnarly as
anything in the guts of Spring Security, and we shouldn't have had to do
it ourselves — not having to do this shit is WHY WE USE FRAMEWORKS IN
THE FIRST PLACE — but here we are nonetheless.