Clean up the testing by not poking OpenID Connect during test
It's imperfect, but the OIDC config is very sensitive to exactly how it
is used during test, causing wild problems when running tests inside
Eclipse. I've no idea why the loading of the truststore isn't happening
at the right time, but it isn't and the pain isn't worth it, especially
when the tests in question aren't supposed to be of the OIDC config!

Also a cleaner way of configuring the truststore, generating it in a
subdir of target that is marked as a place to get (binary) resources
from.