Spring Boot now prohibits circular references by default
This breaks the circularity by making the executor be lazy, but that
means it can't be used during startup or shutdown of the DB connector.
Fortunately, that's also when we don't really need it; the init and
cleanup SQL are allowed to run as long as they need. See
spring-projects/spring-boot#27652 for the trigger

Also, by wrapping it we can return a constant future in the cases where
we can't actually schedule anything, making the rest of the transaction
handling a bit neater. Which is nice.