Workaround for spring-projects/spring-framework#19456
This is really ugly, and there's no mention of this landmine in the docs
so it's a nasty surprise. Not 100% surprising when looked at as "this is
a way to call this method with these arguments", but horrible when
dealing with the use-case where you're generating URLs to put into form
handlers.

At least .query(null) is easy to do.