Group membership synch implementation
If there's a better way than loading the stuff into a temp table, I
don't know it. Using a temp table is easy, but requires a bit of a dance
because the queries that use it won't even compile until after the table
exists. (Adding memberships wouldn't need it, but removing them
definitely does as it needs a view of the whole set at once!)