Timeouts should never get multiplied by 1000
This was a hold-over from when the code was converted from Python, where
we were trying to use the same type of timeout type as there. It didn't
work out and wasn't necessary, but this held on and was a bit more
reachable than expected.

A few other type and doc corrections/improvements too

Conflicts:
	SpiNNaker-comms/src/main/java/uk/ac/manchester/spinnaker/connections/UDPConnection.java