Added FIXME for handling unreachable hardware
I don't know why hardware becomes unreachable, but it's fatal if it is
so, and the recovery strategy has to both mark the hardware as bad
(probably with some kind of report logged) AND ask for the job to be
reallocated anew. This will be messy to fix, because it involves
pushback from the BMP manager to the allocator.