Use Java 11
This PR is about switching to using Java 11 as the baseline Java
version, as 8 is out of official support. (Alas, that doesn't mean we
can switch just yet; that depends on the version of the JDK we have
deployed in various places.)

There will need to be more work to support the additional things the
language has.