Make quota management more forgiving
Without this, when we turn on the new API we're going to get lots of
people suddenly unable to submit jobs because they'll not understand the
importance of specifying which group they're submitting their job
within. Ideally they'd specify the group, but knowing users that's
totally laughable: they won't. So instead, if they don't specify which
group, we get all their groups and pick the one with most quota left.