About as working as I can make things for now
I believe the issue in spring-projects/spring-framework#21515 affects
this code, and it means that I can't yet recommend adopting this PR as
we currently fail all sorts of tests because classpath scanning just
collapses completely after module-isation. The workaround is to run
tests with the module path disabled, but at least Eclipse seems
completely unable to do anything useful with that; you get a total
failure to run tests. This is an abysmal mess!