Quarantine the keycloak client
It's just too nasty for Javadoc, with it claiming that the code uses
modules when it doesn't and I don't understand why. That blocks
publishing the docs, but since it is just for a tool that's really of
not much use to anyone outside of very special circumstances, I don't
mind that.

Also add a few more doc comments to it, since this might be the last
time they get touched...