Login with bearer token
I don't know if this works. I know we want to establish an authenticated
session, and I know that I'm performing an action with a session where I
provide the bearer token at the point when I would otherwise supply the
username and password, but I really don't know if the other side will
like this at all. It might need a different URL to do the proper
exchange of long-term credentials for a session token...