Remove MustBeClosed from progress bars; too much pain down the stack
The problem is that there isn't a way to transfer responsibility for
closing to another (also annotated) object, and I also had to write
constructors just so that the annotation could be added. Bleah. It gets
even worse with SQL connections and statements.

But add them to the NoDropPacketContext; that puts the machine in a
sensitive state while active so it is actually important to close it;
mere garbage collection of it will not do the right thing.