Validate heights and widths
Discovered doing this that spalloc_server's dimensions request was in
triads. That's GREAT because it means that I don't need to measure
rectangles in boards (which don't fit together in a rectangular way).
Chips and triads are much more amenable to validation.

Triad dimensions are limited to 21; 21*21*3*48*18 = 1143072 which should
be enough for the million core machine, and 21*12 = 252 so anything
larger won't be addressable using SDP headers (since we don't want to
ever work with half boards).