Sometimes attaching the tombstone DB fails; that's normally non-fatal
It's not good, but there's a chance things will be fine eventually. The
data isn't lost! It just leaves one of the threads in the pool in an
imperfect state.

Handle this by making a mount failure non-fatal in the first place, and
detecting that this occurred and providing a property for code to see
that it happened and that it shouldn't prod there (which is a single
place in the code, plus tests).