Add locking around connection initialisation
Connection initialisation requires an exclusive transaction, so it isn't
safe to initialise a connection at the same time as anything else, and
there are multiple thread pools where this may be happening so it isn't
practical to simply init the pool ahead of time. Instead, we have to
have a special lock. Which makes sense.