Correct use of GROUP BY
This makes the code use GROUP BY and aggregates correctly, not relying
on the DB to select a special row from the group for us. We simply can't
be sure that the DB will choose correctly. Fortunately, we have indices
already set up that let us efficiently handle this by using the
coordinates of the root of the group (which we were grouping by) and
inflating that back to the ID of the board at that point.

Fixes #693