Pull whether there is a wrap from the DB
I think I've got the queries correct. They're pretty horrible as they
have to do exists checks on a non-trivial join! They might not be
correct for very small machines, but the correct queries will be even
worse, to the point where it'll probably be necessary to formulate them
as some sort of materialized view (the rate at which they change fits
that), and I have no idea how to do that level of wizardry in MySQL...