Reduce timeout on cached HTTP connections
The underlying problem is that Azure is a fairly hostile network
environment by comparison with elsewhere; silent failures of sockets
left open are extremely likely. This matters especially with the codeql
analysis builds whenever the POM was updated (because that flushes
caches and is a somewhat longer build than others use). If we
preemptively close long-inactive sockets earlier, we don't hit the
failure mode.

See https://stackoverflow.com/q/55899091/301832 for details and know
that we've been seeing these occasionally.