Set job timeouts
Sets timeouts for various Github Actions jobs. In particular, the
timeouts are set so that normal running times will be quite a bit
shorter than the timeouts, yet the timeouts won't be the hugely long
default ones (360 minutes!).

Also, delete a job that was unlikely to work (probably broken trigger)
and wasn't important to us.