Do not fail when non-critical workflow steps fail
Some workflow stages are non-critical. They can fail for various reasons
(such as complexities of credentials for bots) and yet aren't actually
very important if they don't run successfully for a particular workflow
job. This marks them as something that the workflow can keep going
after. (Coverage reporting already was such a stage.)

This was causing dependabot-submitted jobs to fail.